Updated openstack/openstack Project: openstack/tripleo-heat-templates  1b237fa9a7797a2247c24b0387a28d65e97e1f4e
Output internal API VIPs

Seeding of overcloud keystone endpoints is currently done via a script
that is external to the overcloud heat stack. Previously the script
didn't have a way to figure out what are the IP addresses that it should
use for internal service endpoints. This patch adds those IP addresses
into the stack outputs so that the script can properly configure
internal endpoints.

Change-Id: I9ae4fc4413a79d6b7e2dce1571fd7083c23348ca