Updated openstack/openstack Project: openstack/barbican  db5c8054659605199a36b078734e31dd912fb75a
Fix unit test errors caused by new mock version

This patch corrects a few errors that is currently breaking the gate
checks.  The trigger of these errors is an upgrade to Mock, but the
root is subtle issues with how Mock was being used.  All of these
changes will work correctly with the old and new versions of Mock.

1) Fix missing attribute errors (calls to non existing assert checks)
2) Fix use of assert_has_calls() (parameter should be a list)
3) Don't assume name of Mock instance
4) Upgrade to use of newer PBR to address import errors
       (this change will be forthcoming from an infra-bot)

Co-authored-by: Kaitlin Farr <kaitlin.farr@jhuapl.edu>
Change-Id: Id4f46ae77bda5a89fb112a779c089542c3987e7b