Updated openstack/openstack Project: openstack-infra/system-config  dd2acc1467d3aa1fcdac7022f036974b9f1e9511
Add puppet-community/archive module

The puppet-community/archive module models well the 'wget file, unzip'
pattern. With it, we can formalize our pattern for doing that. Originally was
nanliu/archive.

An example of where this could be used is the 'download-zanata' exec in
the zanata module. It uses wget right now.

http://git.openstack.org/cgit/openstack-infra/puppet-zanata/tree/manifests/init.pp#n74

Change-Id: I002951a492ef25d86b3bfd5df01429edd3da08fd