Updated openstack/openstack Project: openstack-infra/system-config  12f360bd4a2817bda514ba6a354138d4b3d0e752
Use ubuntu-trusty for multi-node testing

cloud-init does what you'd think we want for the hostname, except it's
not because our hostnames are long. ubuntu-trusty nodes used
ubuntu-minimal and simple-init/glean to process things. Add a 2-node
profile based on ubuntu-minimal so that we can use that. A follow up
patch will remove the old devstack-minimal.

Change-Id: I67ea62ebf9360f72f121d77acd1cbd35d8db728b