Updated openstack/openstack Project: openstack-infra/subunit2sql  a3288c954a2304b8b52f195923d2900c3f889625
Update the run_time graph image in docs

This commit updates the image used in the subunit2sql-graph docs for
the run_time command. There was a fix pushed to set a floor value of
0 for the std_dev backfill since the docs were first added. This
updates the graph to use one that has this fix applied. This however
comes with the tradeoff of showing the few weeks were the infra
subunit2sql gearman worker was down and no data was collected.

Change-Id: I797db1d8a245cd6a89155ff716e09c61275aa2cf