Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  50f66b2d8606dcca712de29ba861901b0b8bc967
Add deprication warning in documentation

According to the message on line 443:

branch-compare-type and branch-pattern at project
level are deprecated and support will be removed
in a later version of Jenkins Job Builder;

This has the potential to create some confusion.

I have flagged the branch-pattern and branch-compare-type documentation
to show that they have been depricated and added a pointer to the
proper alternative (branches).

Change-Id: Id732106aefedc95d37e860e60003459539008d25