Updated openstack/openstack Project: openstack/neutron  4895cac7bd88555fbb49287b57744cd4cdbd6914
Install more-specific ICMPv6 rule in DVR routers

The Openflow rule added in install_dvr_process_ipv6() is dropping
all ICMPv6 traffic, not just the Router Advertisement the comment
mentions. This is causing things like ping6 to fail to VMs on DVR
compute nodes because the reply packets are getting dropped in the
local DVR router before being sent to br-tun.

Change-Id: I14741dd4293e7cfb99cb6bba03cf583ca3ea82ef
Closes-bug: 1471000