Updated openstack/openstack Project: openstack/oslo.versionedobjects  93100cfe02cef77c64263701ddad7fa660300cf6
Compound fields now coerce their elements' values

Added a special datastructures (CoercedList, CoercedDict and CoercedSet)
which are able to coerce their values to appropriate element type.
The instances of these structures are returned when the values of
appropriate compound fields initially coerced. As a result, all
subsequent calls which are going to add new elements to the fields'
collections will be processed by the overridden methods of new
datastructures, thus the new elements will be coerced and the exceptions
risen if the coercing is not possible.
This also fixes a bug in TestObjMakeList which was specifying incorrect
obj_name for the elements in 'objects' collection of MyList test class.
It was not leading to failures as the items of this collection were not
being coerced  prior to this patch.

This patch restores the functionality of patch
2da963c3b313e14c4e60301e87f9c9260d7166cb which was reverted due
to found bug. This one fixes this bug by removing custom initializers for
coerced collections.

Change-Id: I9dc8fc408cefea921e4e27c0e75f98291674f867
Partial-Bug: #1469865