Updated openstack/openstack Project: openstack/sahara-image-elements  cd2dbdcdd3a74f7c1747230ce59579322dd8b944
openjdk: revert to a semi-manual installation

Unfortunately, using package-installs means there is no way to set the
proper openjdk package depending on the version of the distro, and so
far it is hardcoded to 7. On Fedora >= 21, the default openjdk is 8, and
due to bugs in yum [1] this has not caused build errors so far, but
leaving a java-less image.

The chosen solution is to not use package-installs anymore, still using
pkg-map (called manually) to give the right name for the built distro.
The default is still set to install openjdk v7 everywhere, as before;
as drawback, building images based on Fedora >= 21 (which is the default
in diskimage-builder since version 0.1.43) will fail hard now.

[1] like not exiting when asking to install non-existing packages if
specified together with existing ones

Change-Id: I53195daa8d29b15dcae548eb76c5c9efb6ad50ef