Updated openstack/openstack Project: openstack/sahara  eb1d949dff9db5e177e0cc3b4fb2a2dba1ac8baf
[EDP][Oozie] external hdfs missed for URLs in job_configs

If a job is run on data sources that reference an hdfs host not
listed in the /etc/hosts file of cluster instances, the Oozie EDP
engine will update the /etc/hosts files accordingly.However, this
only works for hdfs URLs referenced via data sources. If for
example a Java job is run with manually specified input or output
URLs that reference an external hdfs, the Oozie EDP engineu will
not update the /etc/hosts files.

Change-Id: I42a89a347cdb1bd6cc9d8b8c2b03b8a09d2e17b9
Closes-bug: 1425731