Updated openstack/openstack Project: openstack-infra/project-config  734669fa37a4832c72d4200c4af52098e40c2f5d
Add project fuel-agent repo

This repository is for Fuel Agent (provisioning tool for Fuel).
We used to keep it https://github.com/stackforge/fuel-web/fuel_agent.
Now we need it to be placed in a separate repository because
we want it to be independent component with its own core
team and own release cycle.

Change-Id: I0b183171839557719544fa59acd35b284f49eb2f