Updated openstack/openstack Project: openstack/diskimage-builder  cdf4467a8a604d26f3235bb093afbf008fa1d68f
Cleanup the build directories earlier

Split the cleanup_dirs function in two, i.e. cleanup of the build dir
and the image dir, and use the former to cleanup the temporary build
subdirs after their unmount, before the conversion to other disk
formats; they are not needed anyway at that point, and allows to save
disk space during the conversion phase.

Change-Id: Ie30d7e6033613d6979148423326ae7e17a7342e7