Updated openstack/openstack Project: openstack/neutron-specs  5d465a7af761bc58f896f66257ca0533c25e9bb7
Propose scripted online schema migrations

This specification discusses the issue of database schema
migrations which may proceed while allowing both the previous
and the updated version of the Neutron database API to run
against that schema at the same time.  This is part of a
larger approach which is to allow a Neutron application
to be upgraded to a new version without incurring downtime
while the database schema is migrated.

Change-Id: Ib27c6eb8c901536f1f30d0911dedd4db13d76e9f