Updated openstack/openstack Project: openstack/nova  0c7d5c44f6ca3625be3a2bb458acf137a15102b1
Make NoValidHost exceptions clearer

Adds the MaxRetriesExceeded exception class as a subclass of
NoValidHost, and changes the exceptions raised in those cases to the
more specific MaxRetriesExceeded. Also added an informative comment to a
previously empty NoValidHost exception in the spirit of improving the
reporting of these failures.

Change-Id: I286a215504ed5af8779d6a4d9249842633f47f50
Partially-Implements: blueprint: no-valid-host-reporting