Updated openstack/openstack Project: openstack-infra/project-config  d3dfbb8f881a19152d7275447b90f01a4a28587b
Add excludes support for create-zanata-xml

For projects such as Manuals and Horizon, the source directory in
zanata.xml is set to '.', which may have the unintended side-effect
of the Zanata client finding .po or .pot files under dot directories.
Add a new option to create-zanata-xml, and use it for those two
projects.

Change-Id: I5fc26069c288eca130a45c6e9c5e3dd00ad2c784