Updated openstack/openstack Project: openstack/tripleo-incubator  6e1cf20af5dd3f4787438be605bc0b5b2579cc27
Make sure configure-vm mac doesn't start negative

The original version of this code allowed for the possibility of
generating macs with negative numbers (consider generating 100 macs,
with a random start value of 0xaa).

While this is unlikely to be a problem in real-world uses of this
code, it's still a potential bug if someone starts to push on the
limits of what is supported.

Change-Id: Ia42da0ece5ab3fbeccd6b4d8a4125c1bf5a658d0