Updated openstack/openstack Project: openstack/nova  79a3b1d983b93b9875109919b064310aa67a7568
make test_save_updates_numa_topology stable across python versions

test_save_updates_numa_topology has a rather complex nova object to
represent numa. This includes items which are sets, but must be
converted to lists for json. This is not a stable output operation
across all space and time. The mock comparator was failing here some
times because of it.

Instead, we can manually inspect the parameter in question with our
json comparator to ensure that it is correct.

Closes-Bug: #1473949

Change-Id: Ibfa441356d10417e2b1e092826562ad03d8cfd02