Updated openstack/openstack Project: openstack/oslo.messaging  8a9b5d44ae8d59591dd0e7a8e75bb02ade4b02d7
Executor docstring & attribute tweaks

Adjust some docstrings to better align with the class
they are attached to. Also changes some docstrings for
the pooled executor (as it is an abstraction on-top of
async executors, and is not tied to threads anymore) and
correctly denote this on its attributes that others can
and are overriding. 

Change-Id: I2eaaeed3aa0ae64eb5c82843381b7518ac508ac1