Updated openstack/openstack Project: openstack-infra/puppet-bugdaystats  0404830aca658422b5fa449a9ecdf1df364b5fbb
Add Gemfile and puppet 4 checks

In anticipation of puppet 4, start trying to deal with puppet 4 things
that can be helpfully predicted by puppet lint plugins.

This patch also corrects errors caught by the
puppet-lint-empty_string-check gem. It makes the
$bugdaystats::site::configfile and $bugdaystats::site::httproot
parameters required since the file resources named after them could not
possibly work if they were left undefined or empty string. The
parameter $bugdaystats::site::serveradmin is currently unused in this
module, so we make it optional and default to undef in preparation for
when this functionality is added.

Change-Id: I9eef9a8d49c1ca597783b19fbbbe0129c6157cd5