Updated openstack/openstack Project: openstack/heat  8fe749e8a052208ec8d61ca5b8254a09e64622db
Move signals to SignalResponder class

This change relocates the code that creates and deletes heat, swift and
zaqar signals out of SoftwareDeployment and into SignalResponder, as a
first step in making these signals available to all resources. As part of
this change the structure of the Swift URLs was changed to match that of
the SwiftSignal resource.

Change-Id: Id5701b3696b0ab41433a3f158fabce1f36aabe16
Implements: blueprint uniform-resource-signals