Updated openstack/openstack Project: openstack/governance  b8d930c3a155d2a59311b1f8009129b7e87ef38b
No longer support attributes in tags

This postulates that tags should be tags, and therefore shouldn't
have attributes.

Attributes in tags were originally meant to hold additional
information on the area described by the tag. However, ideally
tags are self-contained units of information, with an opinionated
description (and ideally an objective set of application rules).
Keeping the option of having attributes in tags adds useless format
complexity and steers tags in the direction of structured project
metadata. Project structured metadata (like the ops-data) is fine,
it's just not the same thing as tags. Continuing to support
attributes in tags IMHO encourages the confusion between the two.

The integrated-release tag was the only one using attributes,
and it's now removed.

Change-Id: I1dd5c8090405bd87e8947892d78c62509258e363