Updated openstack/openstack Project: openstack/python-ironicclient  8e23e8add406f16e1d0b7df129e0917600b1cfbe
Allow specifying a set of fields of the Node resource

This patch add a "--fields" parameter to the "node-list", "node-show"
and "node-port-list" commands which the user can specify a subset of
fields that will be returned by the server.

This is supported by the Ironic API version >= 1.8, so this patch also
bumps the default API version to 1.8.

Related-Bug: #1466495
Change-Id: I40654ee9fbd92dd91b41f8596adcd26264634147