Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  ab89a02fe9cfe128952cc38a7a9c69df6becf6bf
Validate and set compare types in gerrit trigger

Gerrit Trigger does not validate compare types for projects, branches,
file paths, and topics currently and this is error prone.

This change adds a new function which validates the specified compare
types against the list of valid ones and returns the compare type if it
is valid. If not, it throws InvalidAttributeError.

Change-Id: Ic9c17fab2baf3a4b0bd06c3cda319ea7cbf9d623
Signed-off-by: Fatih Degirmenci <fatih.degirmenci@ericsson.com>