Updated openstack/openstack Project: openstack/horizon  4fa3abce3b5d1727fd22b1a9b1cd5934db10189b
Revert "Ignore the ng_on_watch error from eslint"

This reverts commit 9b210766508bfeab6cb7f1c4d0c1524a51209613.

The original commit was merged with a -1. The concern raised was
that it does not force engineers to clean up after themselves, which
is a best practice explicitly mentioned in the John Papa style guides
under Y070. While this rule is only specific to directives, it's
common sense, and the directive to "clean up your pointers" goes
back as far as C.

https://github.com/johnpapa/angular-styleguide#style-y070

Change-Id: Iea243c2b0b3c6154c5b02fd9e29172e1d5e3b434