Updated openstack/openstack Project: openstack/cinder  c84976fd3d0dd453878faf2d8842ef187bb037f9
Dell SC: Add support for ManageableVD

Added support for the ManageableVD volume driver class. This also
meant the driver changed from inheriting the volumedriver class to
inheriting the assorted *VD classes.

A few things about our ManageableVD support. We only allow management
of a volume if it is sized in multiples of 1GB. We only allow
management of volumes that are not currently mapped to a server on the
Storage Center. id-types of source-id (DeviceId on the Storage Center)
and source-name (Volume Name on the Storage Center) are supported.

Change-Id: Iecc93ce1654111e1292cfc1931d1eda4abd08595