Updated openstack/openstack Project: openstack-infra/shade  a878fb41fa74d1dad8b6ab64ff9a78e0a366fa7d
Account for Error 396 on Rackspace

Error 396 doesn't have a name, or a description. It's just Error 396. It
shouldn't be judged for its failings in being descriptive, nor for the
utter lack of necessity of its existence. Instead, we should feel sorry
for it and the meaningless life it lives. You see, Error 396 is the
error that the asynchronous image importing code returns when an unknown
error is encountered and retrying is the appropriate action to take. Why
doesn't the importing worker simply retry itself? Only Cthulhu can
possibly know the answer to such a dark question, but perhaps it's for
the best. Even a meaningless life such as the one Error 396 leads is
better than no life at all. Or?

Change-Id: I848986c9616d96caaad69806ad8b4f9f10301768