Updated openstack/openstack Project: openstack/requirements  6a77e06d1b49f5a7a9d768209cc1e09b4dde8db8
bump keystonemiddleware to g-r compatible version

The current minimum of keystonemiddleware has a pbr requirement of
pbr (>=0.6,!=0.7,<1.0) this forces a downgrade of pbr to 0.11 which
is incompatible with the current requirements and syntax of
global-requirements. Version 2.0.0 of keystonemiddleware has a pbr
requirement of pbr (>=0.11,<2.0) which will not force a downgrade of
pbr.

Change-Id: I5c7f8bc6fcb495e6866d8405effbff78ff0179c3