Updated openstack/openstack Project: openstack-infra/project-config  23bb9a7ac01ef726cd01a36d0fe500b857e22429
Drop unstable configuration for tempest-dsvm-neutron job

The introduction of pymysql triggered race conditions in the Neutron's IP management
layer. In order to address these without disrupting ongoing work and the gate, a
number of infra-related changes [2] were submitted to allow the team to triage and
experiment with possible solutions.

Since then, we've been fixing those races and watching the failure rates of both
stable and unstable configurations. They now look remarkably close [3], thus giving
us confidence that this configuration is no longer necessary.

This patch removes the job, whereas [4] removes the devstack-gate knobs.

The occasional skew between the two curves is due to the fact that the full job
runs a lot more often (beyond just Neutron changes) compared to the unstable one
(configured to run on Neutron changes only), thus affecting the mobile average.
Therefore, this is not a sympthom of the fact that there are still issue lurking
in the dark, at least I until I am proved wrong!

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-June/066695.html
[2] https://review.openstack.org/#/q/topic:neutron-unstable,n,z
[3] http://goo.gl/YM7gUC
[4] https://review.openstack.org/#/c/199672/

Change-Id: I886f64a2379a8e16957e6ca4bcf784cb22ad2c28