Updated openstack/openstack Project: openstack-infra/tripleo-ci  7f86e1e38db206d67147f4b4131bd3d622221e1c
Disable MongoDB journaling in the CI

When MongoDB journaling is enabled, MongoDB will try to create 3 1G file
on the filesystem in the datapath (One for journal j_, and 2
preallocated prealloc.{1,2}). Those files are usefull for production
needs when 'data integrity' is needed. Since CI instance are ephemeral
this feature is not needed in CI.

Change-Id: I214abbda3c2f2556c67035611cac2347d6ef00fa
Depends-On: I0e4e65af9f650c10fdf5155ff709b4eb984cf4e1