Updated openstack/openstack Project: openstack/ironic-python-agent  d71258b1ea7e36fe00e747f7bf7f421f36563750
Change Dockerfile to use Debian as a base

Ubuntu licensing does not allow for distributing derivatives of Ubuntu,
including containers based on Ubuntu. We currently publish a CoreOS PXE
image with an IPA container embedded. Basing this container on Ubuntu
appears to violate Ubuntu's license. Let's move to Debian to fix this.

Also remove the python3 purge as python3 doesn't exist in the Debian
base image.

Closes-Bug: #1475325
Change-Id: I051e3123f0cd1e66b5e3bae727559fe31467791b