Updated openstack/openstack Project: openstack/neutron-vpnaas  22468ce55d16c1b7add6b9cb4353e72caccbcb16
Scenario test for vpnaas: ipsec-site-connection

The test verifies that it is possible to establish a vpn
connection between two sites. And then it performs basic
connectivity test. It pings IP address of site2 from an
IP address of site1.
The patchset implements scenario test for both
openswan and strongswan device drivers.

Change-Id: I5ff943d4cb4de06c79c2f2d0fee178ad992f36d1