Updated openstack/openstack Project: openstack/nova  c352c43b4a805a0c0b0c749d9c240768b0335aed
Cells: add instance cell registration utility to nova-manage

The patch enables an admin to register instances within a cell.
This is done via:

  nova-manage cell_v2 register_instances

With the following parameters:
     --cell_uuid - mandatory - the cell uuid
     --limit - optional - the maximum number of mappings to be done
                          in this iteration
     --marker - optional - this should be used if the previous run
                           had a limit parameter
     --verbose - optional - prints out useful debug information

NOTE: each time the utility is run it will print the marker that
should be used for the next iteration.

DocImpact

Implements blueprint cells-instance-migration

Change-Id: Ic03fda8bbd0b8d07e61fc08f240750c17c3ef28a